+++ src/components/pages/Login.tsx
@@ -1,30 +1,15 @@
-import {
-  ChangeEvent,
-  FC,
-  memo,
-  useCallback,
-  useContext,
-  useState,
-} from "react";
+import { ChangeEvent, FC, memo, useCallback, useState } from "react";
 import { Box, Flex, Heading, Link, Text } from "@chakra-ui/react";
-import { useNavigate } from "react-router";
-import Cookies from "js-cookie";
 
 import { PrimaryButton } from "../atoms/PrimaryButton";
 import { PrimaryInput } from "../atoms/PrimaryInput";
 import { PasswordInput } from "../molecules/PasswordInput";
-import { SignInParams } from "../../theme/api/userAuth";
-import { signIn } from "../../api/auth";
-import { LoginUserContext } from "../../providers/LoginUserProvider";
-import { useToastMsg } from "../../hooks/useToastMsg";
+import { useSignIn } from "../../hooks/useSignIn";
 
 export const Login: FC = memo(() => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const { showToastMsg } = useToastMsg();
-
-  const { setIsSignedIn, setCurrentUser } = useContext(LoginUserContext);
-  const navigate = useNavigate();
+  const { signIn } = useSignIn();
 
   const onChangeEmail = useCallback(
     (e: ChangeEvent<HTMLInputElement>) => {
@@ -40,26 +25,8 @@ export const Login: FC = memo(() => {
     [password]
   );
 
-  const onClickSignIn = useCallback(async () => {
-    const params: SignInParams = {
-      email,
-      password,
-    };
-
-    try {
-      const res = await signIn(params);
-      if (res.status === 200) {
-        Cookies.set("_access_token", res.headers["access-token"]);
-        Cookies.set("_client", res.headers["client"]);
-        Cookies.set("_uid", res.headers["uid"]);
-        setIsSignedIn(true);
-        setCurrentUser(res.data.data);
-        navigate("/");
-        showToastMsg({ status: "success", title: "ログインしました" });
-      }
-    } catch (err) {
-      showToastMsg({ status: "error", title: "ログインに失敗しました" });
-    }
+  const onClickSignIn = useCallback(() => {
+    signIn({ email, password });
   }, [email, password]);
 
   return (
